Move the code from MVC to React front end

+++ Enrolliks.Web/People/PeopleController.cs
@@ -4,7 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Enrolliks.Web.People
 {
-    public class PeopleController : Controller
+    [ApiController]
+    public class PeopleController : ControllerBase
     {
         private readonly IPeopleRepository _repository;
 
@@ -13,38 +14,48 @@ namespace Enrolliks.Web.People
             _repository = storage;
         }
 
-        [HttpGet]
-        public async Task<IActionResult> Index()
+        [HttpGet("[controller]/list")]
+        public async Task<IActionResult> GetAll()
         {
             var people = await _repository.GetAllAsync();
-            return View(people);
+            return Ok(people);
         }
 
-        [HttpGet]
-        public IActionResult Create()
-        {
-            return View(new CreatePersonModel());
-        }
-
-        [HttpPost]
+        [HttpPost("[controller]/create")]
         public async Task<IActionResult> Create(CreatePersonModel personModel)
         {
             if (string.IsNullOrWhiteSpace(personModel.Name))
             {
-                return RedirectToAction();
+                return StatusCode(400);
+            }
+
+            bool exists = await _repository.ExistsAsync(personModel.Name);
+            if (exists)
+            {
+                return StatusCode(409);
             }
 
             var person = new Person(personModel.Name);
             await _repository.CreateAsync(person);
-
-            return RedirectToAction(nameof(Index));
+            return Ok();
         }
 
-        [HttpPost]
-        public async Task<IActionResult> Delete(string name)
+        [HttpDelete("[controller]/delete")]
+        public async Task<IActionResult> Delete(string? name)
         {
+            if (string.IsNullOrWhiteSpace(name))
+            {
+                return StatusCode(400);
+            }
+
+            bool exists = await _repository.ExistsAsync(name);
+            if (!exists)
+            {
+                return StatusCode(404);
+            }
+
             await _repository.DeleteAsync(name);
-            return RedirectToAction(nameof(Index));
+            return Ok();
         }
     }
 }
